Fixed bad codon problem
Turns out the bad codons were not succesfully deleted in some instances.
Could trace that back to 
for item in self.expaltcodons:
            if item in bad_codons: self.expaltcodons.remove(item)
Turns out if an item is removed from the list during for item in list loop, the next item is actually at the next index position. E.g. if list[2] is removed, list[3] moves up into the spot of list[2] and the loop does not take that into account, but acts on whatever is now in list[3]. That means actually every item after a to-delete-item gets skipped and thus if two bad codons appear after each other in the list, one of them doesn't get deleted.
Fixed that now.